Process focus.js
